+++ CumulusMX/Utils.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
@@ -335,6 +336,32 @@ namespace CumulusMX
 			{
 				return false;
 			}
+		}
+
+		public static string[] SplitCsv(string line)
+		{
+			List<string> result = [];
+			StringBuilder currentStr = new StringBuilder("");
+			bool inQuotes = false;
+			for (int i = 0; i < line.Length; i++) // For each character
+			{
+				if (line[i] == '\"') // Quotes are closing or opening
+					inQuotes = !inQuotes;
+				else if (line[i] == ',') // Comma
+				{
+					if (!inQuotes) // If not in quotes, end of current string, add it to result
+					{
+						result.Add(currentStr.ToString());
+						currentStr.Clear();
+					}
+					else
+						currentStr.Append(line[i]); // If in quotes, just add it
+				}
+				else // Add any other character to current string
+					currentStr.Append(line[i]);
+			}
+			result.Add(currentStr.ToString());
+			return result.ToArray(); // Return array of all strings
 		}
 	}
 }